Rename `data_samples` to `data_test_cases`
